% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_destinations.R
\name{get_destinations}
\alias{get_destinations}
\title{Get destinations}
\usage{
get_destinations()
}
\value{
A data frame of supported destinations with the following
columns:
\itemize{
\item destination_id
\item destination_name
\item destination_slug
\item park_id
\item park_name
}
}
\description{
Get a list of supported destinations available on the live API
}
