% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity}
\alias{get_entity}
\title{Get entity document}
\usage{
get_entity(entity_id)
}
\arguments{
\item{entity_id}{Either the GUID or the slug string for a given entity}
}
\value{
A data frame with details about the given entity
}
\description{
Get entity document
}
\examples{
## Entity data for Walt Disney World
\donttest{
get_entity("waltdisneyworldresort")
}
}
