% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity_children}
\alias{get_entity_children}
\title{Get entity children}
\usage{
get_entity_children(entity_id)
}
\arguments{
\item{entity_id}{Either the GUID or the slug string for a given entity}
}
\value{
A data frame with details about the given entity's children
}
\description{
Get all children for a given entity document
}
\details{
This is recursive, so a destination will return all parks and all rides
within those parks.
}
\examples{
## Entity children data for Walt Disney World
\donttest{
get_entity_children("waltdisneyworldresort")
}
}
