% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity_schedule}
\alias{get_entity_schedule}
\title{Get entity schedule}
\usage{
get_entity_schedule(entity_id, year = NULL, month = NULL)
}
\arguments{
\item{entity_id}{Either the GUID or the slug string for a given entity}

\item{year}{Numeric. Year for entity schedule (if \code{NULL} will default to
current year)}

\item{month}{Numeric. Month for entity schedule (if \code{NULL} will default to
current month)}
}
\value{
A data frame with schedule data for the entity
}
\description{
Get this entity's schedule for the next 30 days
}
\examples{
## Schedule data for Walt Disney World
\donttest{
get_entity_schedule("waltdisneyworldresort")
}
}
