% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get forecast}
\usage{
get_forecast(entity_id)
}
\arguments{
\item{entity_id}{Character. Either the GUID or the slug string for a given
entity.}
}
\value{
A data frame of forecast data with three columns:
\itemize{
\item time
\item wait_time
\item percentage
}
}
\description{
Get forecast
}
