% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wait_times.R
\name{get_wait_times}
\alias{get_wait_times}
\title{Get Wait Times}
\usage{
get_wait_times(entity_id)
}
\arguments{
\item{entity_id}{Character. Either the GUID or the slug string for a given
entity.}
}
\value{
A data frame with three columns:
\itemize{
\item park
\item name
\item wait time
}
}
\description{
Get Wait Times
}
\examples{
## Get wait time for the Jungle Cruise
\donttest{
get_wait_times("junglecruise")
}
}
