% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_calendar.R
\name{park_calendar}
\alias{park_calendar}
\title{Get park calendar}
\usage{
park_calendar(park)
}
\arguments{
\item{park}{Park ID, see \code{parks()}}
}
\description{
Return a list of park information.
}
\examples{
park_calendar("WaltDisneyWorldMagicKingdom")

}
