% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_info.R
\name{park_info}
\alias{park_info}
\title{Get park information}
\usage{
park_info(park)
}
\arguments{
\item{park}{Park ID, see \code{parks()}}
}
\description{
Return a list of park information.
}
\examples{
park_info("WaltDisneyWorldMagicKingdom")

}
