% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_wait_time.R
\name{park_wait_times}
\alias{park_wait_times}
\title{Get park wait times}
\usage{
park_wait_times(park)
}
\arguments{
\item{park}{Park ID, see \code{parks()}}
}
\description{
Return a list of park wait times.
}
\examples{
park_wait_times("WaltDisneyWorldMagicKingdom")

}
