% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeparks-package.R
\docType{package}
\name{themeparks-package}
\alias{themeparks}
\alias{themeparks-package}
\title{themeparks: An Interface to Themeparks.wiki}
\description{
themeparks allows you to interact with theme park data from the
themepark.wiki API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucymcgowan.github.io/themeparks/}
}

}
\author{
\strong{Maintainer}: Lucy D'Agostino McGowan \email{lucydagostino@gmail.com}

}
\keyword{internal}
